Upgrade config and scripts for Postgres 16
Changes to the two Postgres config files (postgresql.conf and
pg_hba.conf) are mostly due to using a Postgres 16 template and copying
over the changes we had made.
Lines in postgresql.conf marked with "#!!!!!" are ones I have customised
for our use-case. All other changes are just those made by the upgrade.

A new config file specifically for loading from backup is added; this
MUST NOT be used in production, as it disables important crash-recovery
functionality that slows down data ingestion.

Promoting a hot standby to primary can no longer be done by a trigger
file, so update the script accordingly.